+++ templates/token-minting-dapp-template/frontend/view-functions/getRegistry.ts
@@ -1,0 +1,12 @@
+import { AccountAddress } from "@aptos-labs/ts-sdk";
+import { aptosClient } from "@/utils/aptosClient";
+import { MODULE_ADDRESS } from "@/constants";
+
+export const getRegistry = async () => {
+  const registry = await aptosClient().view<[[{ inner: string }]]>({
+    payload: {
+      function: `${AccountAddress.from(MODULE_ADDRESS)}::launchpad::get_registry`,
+    },
+  });
+  return registry[0];
+};